moved error alert has a seperate component
