+++ frontend/src/components/auth/LoginForm.tsx
@@ -1,4 +1,4 @@
-import React, { useContext, useRef, useState } from "react";
+import React, { useContext, useState } from "react";
 import {
   Box,
   Button,
@@ -7,8 +7,6 @@ import {
   TextField,
   Typography,
   useTheme,
-  Alert,
-  AlertTitle,
 } from "@mui/material";
 import { AlternateEmail, Password } from "@mui/icons-material";
 import InputAdornment from "@mui/material/InputAdornment";
@@ -24,6 +22,7 @@ import {
   defaultAuthError,
 } from "../../errors/AuthErrors";
 import { capitalizeFirstLetter } from "../../utils/StringUtils";
+import ErrorAlert from "./ErrorAlert";
 
 function isDrawerLogin(userRole: Role) {
   return userRole !== "Learner";
@@ -44,7 +43,6 @@ const LoginForm = ({ userRole }: LoginFormProps) => {
   const [errorData, setErrorData] =
     useState<[string | undefined, string | undefined]>();
 
-  const alertRef = useRef<HTMLDivElement>(null);
   const theme = useTheme();
 
   const onLogInClick = async () => {
@@ -86,47 +84,6 @@ const LoginForm = ({ userRole }: LoginFormProps) => {
           setLoginError(defaultAuthError);
       }
     }
-  };
-
-  interface ErrorAlertProps {
-    title?: string;
-    message?: string;
-  }
-
-  const ErrorAlert: React.FC<ErrorAlertProps> = ({ title, message }) => {
-    return (
-      <Box ref={alertRef} sx={{ width: "100%" }}>
-        <Alert
-          icon={false}
-          severity="error"
-          sx={{
-            color: theme.palette.Error.Light,
-            width: "100%",
-            height: "100%",
-            borderRadius: "4px",
-            border: "2px solid",
-            borderColor: theme.palette.Error.Hover,
-            padding: "20px",
-            display: "flex",
-            flexDirection: "column",
-            alignItems: "flex-start",
-            "& .MuiAlert-message": { padding: 0 },
-          }}
-        >
-          <AlertTitle marginBottom="3px">
-            <Typography
-              variant="titleMedium"
-              color={theme.palette.Error.Default}
-            >
-              {title || "Error"}
-            </Typography>
-          </AlertTitle>
-          <Typography variant="bodyMedium" color={theme.palette.Error.Pressed}>
-            {message || "An error has occurred."}
-          </Typography>
-        </Alert>
-      </Box>
-    );
   };
 
   return (
